requirements.txt
# JiekouAI ComfyUI Plugin Dependencies
# Python 3.10+

requests>=2.28.0
pillow>=9.0.0
numpy>=1.21.0
torchaudio>=2.0.0
opencv-python>=4.7.0

